+++ supperbas/QLJava/src/ast/expression/unary/Neg.java
@@ -15,7 +15,7 @@ public class Neg extends Unary {
 	
 	@Override
 	public Type typeOf(Map<ast.type.Ident, Type> typeEnv) {
-		return new ast.type.Numeric();
+		return new ast.type.Int();
 	}
 
 	@Override
