+++ supperbas/QLJava/src/ast/expression/unary/Pos.java
@@ -15,7 +15,7 @@ public class Pos extends Unary {
 	
 	@Override
 	public Type typeOf(Map<ast.type.Ident, Type> typeEnv) {
-		return new ast.type.Numeric();
+		return new ast.type.Int();
 	}
 
 	@Override
